+++ CHANGELOG.md
@@ -256,6 +256,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Objects inside of stacks will now retain their custom colour after reloading
   a previous save.
 - An object's centre-of-mass will no longer change when the object is scaled.
+- Fixed an issue where if the 'Red Cup' container had a custom scale applied to
+  it, the applied scale would change when loading a previous save state.
 
 #### Multiplayer